feat: add firmware version reporting through full stack
- Create src/version.h with BRAMBLE_FIRMWARE_VERSION constant using
  uint32_t encoding: (major << 24) | (minor << 16) | build
- Change firmware_version from uint16_t to uint32_t across all firmware
  structs (RegistrationPayload, NodeInfo, NodeConfiguration,
  RegistryNodeEntry) and function signatures
- Add firmware_version to hub LIST_NODES UART output for Pi consumption
- Update Python API models and endpoints to parse and expose
  firmware_version as formatted string (e.g. "1.0.0")
- Add firmware version display to dashboard NodeDetail and NodeCard
- Bump hub registry version to 2 (breaking change in struct layout)
- Replace hardcoded 0x0100 with BRAMBLE_FIRMWARE_VERSION in main.cpp
  and irrigation_mode.cpp
- Log firmware version at boot

+++ src/lora/address_manager.cpp
@@ -16,7 +16,7 @@ AddressManager::AddressManager() : next_available_address_(ADDRESS_MIN_NODE)
 }
 
 uint16_t AddressManager::registerNode(uint64_t device_id, uint8_t node_type, uint8_t capabilities,
-                                      uint16_t firmware_version, const char *device_name)
+                                      uint32_t firmware_version, const char *device_name)
 {
     // Input validation
     if (device_id == 0) {
